QR Cleanup
- Removes QR Code from main file and imports code from qr_generator.py
